added pkg build/extract big files with progressbar

+++ MGZ/source/paged_file.h
@@ -61,7 +61,7 @@ int paged_file_read (PagedFile *f, void *buffer, u32 len);
 int paged_file_getline (PagedFile *f, char **line, int *line_len);
 int paged_file_seek (PagedFile *f, u64 offset);
 int paged_file_write (PagedFile *f, void *buffer, u32 len);
-int paged_file_splice (PagedFile *f, PagedFile *from, int len);
+u64 paged_file_splice (PagedFile *f, PagedFile *from, u64 len);
 int paged_file_flush (PagedFile *f);
 void paged_file_free (PagedFile *f);
 void paged_file_close (PagedFile *f);
